Enableing deploy/destroy commands to be compatible with local deployments

+++ qhub/destroy.py
@@ -34,5 +34,8 @@ def destroy_configuration(config):
             run(["terraform", "destroy", "-auto-approve"])
 
         # 06 Remove terraform backend remote state bucket
-        with change_directory("terraform-state"):
-            run(["terraform", "destroy", "-auto-approve"])
+        # backwards compatible with `qhub-config.yaml` which
+        # don't have `terraform_state` key
+        if config.get('terraform_state') != 'local':
+            with change_directory("terraform-state"):
+                run(["terraform", "destroy", "-auto-approve"])
